refactor: move `Engine` into its own module.
